Fix geotoolkit repo not responding
It causes the whole compilation to break,
sbt doesn't look in another repo.
It seems to be related to https://github.com/sbt/sbt/issues/1138
but workaround is not working.
Also try to improve travis caching as advised in documentation
(geotoolkit repo should not be used if there were cache)